Dedupe, split, and harden token-utils tests further
- Fix jscpd: extract a runRateLimited helper for the pending-work-scoped
  rate-limit calls and an expectPassLookupNotFound helper for the 404
  lookup assertions, removing the duplicated blocks.
- Split the parser/route/cache-header cases into a companion
  token-utils-parsing.test.ts so both files stay under the 400-line test
  limit (faster, narrower mutation runs).
- Guard the success-clears-failures rate-limit test for MAX_TOKEN_404S=1,
  where there is no below-limit state to clear.

Mutation score stays 100% (79/79, 1 suppressed); cpd clean.
